pyproject.toml
[project]
name = "fast-proxy"
version = "0.1.0"
description = "A production-ready FastAPI plugin for transparent HTTP proxying with per-route circuit breakers and OpenTelemetry observability."
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "Satyam Soni", email = "satyam_soni1@epam.com" }
]
requires-python = ">=3.13"
keywords = ["fastapi", "proxy", "http-proxy", "circuit-breaker", "opentelemetry", "resilience"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: Proxy",
    "Topic :: System :: Networking",
]
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-instrumentation-httpx>=0.61b0",
    "opentelemetry-sdk>=1.40.0",
    "pybreaker>=1.4.1",
    "structlog>=24.0.0",
]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpserver>=1.1.5",
]
